clean up reflection patches
